First release
This is the first release, containing weatherData.py and keys.env respectively. keys.env must be filled out with your own data, to avoid exposing an API key to the wide web. weatherData.py is simply run, and has hard-coded values - but it could easily be made more user friendly and customizable.